+++ test/unittest/backends/ibis/__init__.py
@@ -2,7 +2,7 @@ import os
 
 import pytest
 
-skip = not os.environ.get('SUPERDUPER_CONFIG', "").endswith('mongodb.yaml')
+skip = not os.environ.get('SUPERDUPER_CONFIG', "").endswith('ibis.yaml')
 
 if skip:
     pytest.skip("Skipping this file for now", allow_module_level=True)